requirements.txt
customtkinter
pyserial